Moved AsVariantDescs to its own module
